extracted netty server code also reduced dependencies on GeodeRedisServer
