+++ geode-redis/src/main/java/org/apache/geode/redis/internal/GeodeRedisServer.java
@@ -14,53 +14,18 @@
  */
 package org.apache.geode.redis.internal;
 
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.net.InetAddress;
-import java.net.InetSocketAddress;
-import java.net.UnknownHostException;
-import java.security.KeyStore;
-import java.util.concurrent.ThreadFactory;
-
-import javax.net.ssl.KeyManagerFactory;
-
-import io.netty.bootstrap.ServerBootstrap;
-import io.netty.buffer.PooledByteBufAllocator;
-import io.netty.channel.Channel;
-import io.netty.channel.ChannelFuture;
-import io.netty.channel.ChannelInitializer;
-import io.netty.channel.ChannelOption;
-import io.netty.channel.ChannelPipeline;
-import io.netty.channel.EventLoopGroup;
-import io.netty.channel.ServerChannel;
-import io.netty.channel.nio.NioEventLoopGroup;
-import io.netty.channel.socket.SocketChannel;
-import io.netty.channel.socket.nio.NioServerSocketChannel;
-import io.netty.handler.ssl.SslContext;
-import io.netty.handler.ssl.SslContextBuilder;
-import io.netty.handler.timeout.WriteTimeoutHandler;
-import io.netty.util.concurrent.Future;
 import org.apache.logging.log4j.Logger;
 
 import org.apache.geode.cache.Cache;
 import org.apache.geode.cache.Region;
-import org.apache.geode.distributed.internal.InternalDistributedSystem;
-import org.apache.geode.internal.admin.SSLConfig;
 import org.apache.geode.internal.cache.InternalCache;
-import org.apache.geode.internal.inet.LocalHostUtil;
-import org.apache.geode.internal.net.SSLConfigurationFactory;
-import org.apache.geode.internal.security.SecurableCommunicationChannel;
-import org.apache.geode.logging.internal.executors.LoggingThreadFactory;
 import org.apache.geode.logging.internal.log4j.api.LogService;
-import org.apache.geode.management.ManagementException;
 import org.apache.geode.redis.internal.executor.CommandFunction;
 import org.apache.geode.redis.internal.executor.StripedExecutor;
 import org.apache.geode.redis.internal.executor.SynchronizedStripedExecutor;
 import org.apache.geode.redis.internal.executor.key.RenameFunction;
 import org.apache.geode.redis.internal.gfsh.RedisCommandFunction;
-import org.apache.geode.redis.internal.netty.ByteToCommandDecoder;
-import org.apache.geode.redis.internal.netty.Coder;
-import org.apache.geode.redis.internal.netty.ExecutionHandlerContext;
+import org.apache.geode.redis.internal.netty.NettyRedisServer;
 import org.apache.geode.redis.internal.pubsub.PubSub;
 import org.apache.geode.redis.internal.pubsub.PubSubImpl;
 import org.apache.geode.redis.internal.pubsub.Subscriptions;
@@ -78,15 +43,8 @@ public class GeodeRedisServer {
    */
   public static final int DEFAULT_REDIS_SERVER_PORT = 6379;
 
-  private static final int RANDOM_PORT_INDICATOR = 0;
-
   public static final String ENABLE_REDIS_UNSUPPORTED_COMMANDS_PARAM =
       "enable-redis-unsupported-commands";
-
-  /**
-   * Connection timeout in milliseconds
-   */
-  private static final int connectTimeoutMillis = 1000;
 
   private static final Logger logger = LogService.getLogger();
 
@@ -94,82 +52,17 @@ public class GeodeRedisServer {
   private final boolean ENABLE_REDIS_UNSUPPORTED_COMMANDS =
       Boolean.getBoolean(ENABLE_REDIS_UNSUPPORTED_COMMANDS_PARAM);
 
-  /**
-   * The number of threads that will work on handling requests
-   */
-  private final int numWorkerThreads;
-
-  /**
-   * The number of threads that will work socket selectors
-   */
-  private final int numSelectorThreads;
-
   private final PassiveExpirationManager passiveExpirationManager;
 
-  /**
-   * The actual port being used by the server
-   */
-  private int serverPort;
-
-  /**
-   * The address to bind to
-   */
-  private final String bindAddress;
-
-  /**
-   * Temporary constant whether to use old single thread per connection model for worker group
-   */
-  private boolean singleThreadPerConnection;
-
-  /**
-   * The cache instance pointer on this vm
-   */
-  private InternalCache cache;
-
-  /**
-   * Channel to be closed when shutting down
-   */
-  private Channel serverChannel;
+  private final NettyRedisServer nettyRedisServer;
 
   private final RegionProvider regionProvider;
+  private final PubSub pubSub;
 
-  private EventLoopGroup bossGroup;
-
-  private EventLoopGroup workerGroup;
-  private EventLoopGroup subscriberGroup;
-
-  /**
-   * System property name that can be used to set the number of threads to be used by the
-   * GeodeRedisServer
-   */
-  public static final String NUM_THREADS_SYS_PROP_NAME = "gemfireredis.numthreads";
+  private boolean started;
 
   private boolean shutdown;
 
-  private boolean started;
-  private final PubSub pubSub;
-
-
-  /**
-   * Helper method to set the number of worker threads
-   *
-   * @return If the System property {@value #NUM_THREADS_SYS_PROP_NAME} is set then that number is
-   *         used, otherwise {@link Runtime#availableProcessors()}.
-   */
-  private int setNumWorkerThreads() {
-    String prop = System.getProperty(NUM_THREADS_SYS_PROP_NAME);
-    int defaultThreads = Runtime.getRuntime().availableProcessors();
-    if (prop == null || prop.isEmpty()) {
-      return defaultThreads;
-    }
-    int threads;
-    try {
-      threads = Integer.parseInt(prop);
-    } catch (NumberFormatException e) {
-      return defaultThreads;
-    }
-    return threads;
-  }
 
   /**
    * Constructor for {@code GeodeRedisServer} that will configure the server to bind to the given
@@ -182,18 +75,6 @@ public class GeodeRedisServer {
    *        {@value #RANDOM_PORT_INDICATOR} a random port is assigned.
    */
   public GeodeRedisServer(String bindAddress, int port, InternalCache cache) {
-    if (port < RANDOM_PORT_INDICATOR) {
-      throw new IllegalArgumentException("Redis port cannot be less than 0");
-    }
-    this.cache = cache;
-    serverPort = port;
-    this.bindAddress = bindAddress;
-    numWorkerThreads = setNumWorkerThreads();
-    singleThreadPerConnection = numWorkerThreads == 0;
-    numSelectorThreads = 1;
-    shutdown = false;
-    started = false;
-
     if (ENABLE_REDIS_UNSUPPORTED_COMMANDS) {
       logUnsupportedCommandWarning();
     }
@@ -207,6 +88,9 @@ public class GeodeRedisServer {
     RenameFunction.register(stripedExecutor);
     RedisCommandFunction.register();
     passiveExpirationManager = new PassiveExpirationManager(regionProvider.getDataRegion());
+    nettyRedisServer = new NettyRedisServer(() -> cache.getInternalDistributedSystem().getConfig(),
+        regionProvider, pubSub,
+        this::allowUnsupportedCommands, this::shutdown, port, bindAddress);
   }
 
   public void setAllowUnsupportedCommands(boolean allowUnsupportedCommands) {
@@ -230,18 +114,12 @@ public class GeodeRedisServer {
     logger.warn("Unsupported commands enabled. Unsupported commands have not been fully tested.");
   }
 
-  /**
-   * Helper method to get the host name to bind to
-   *
-   * @return The InetAddress to bind to
-   */
-  private InetAddress getBindAddress() throws UnknownHostException {
-    if (bindAddress == null || bindAddress.isEmpty()
-        || bindAddress.equals(LocalHostUtil.getAnyLocalAddress().getHostAddress())) {
-      return LocalHostUtil.getAnyLocalAddress();
-    } else {
-      return InetAddress.getByName(bindAddress);
-    }
+  public RegionProvider getRegionProvider() {
+    return regionProvider;
+  }
+
+  public int getPort() {
+    return nettyRedisServer.getPort();
   }
 
   /**
@@ -249,160 +127,10 @@ public class GeodeRedisServer {
    */
   public synchronized void start() {
     if (!started) {
-      try {
-        startRedisServer();
-      } catch (IOException | InterruptedException e) {
-        throw new ManagementException("Could not start Server", e);
-      }
+      nettyRedisServer.start();
+      passiveExpirationManager.start();
       started = true;
     }
-  }
-
-  public RegionProvider getRegionProvider() {
-    return regionProvider;
-  }
-
-  public PubSub getPubSub() {
-    return pubSub;
-  }
-
-  public EventLoopGroup getSubscriberGroup() {
-    return subscriberGroup;
-  }
-
-  /**
-   * Helper method to start the server listening for connections. The server is bound to the port
-   * specified by {@link GeodeRedisServer#serverPort}
-   */
-  private void startRedisServer() throws IOException, InterruptedException {
-    Class<? extends ServerChannel> socketClass = initializeEventLoopGroups();
-
-    InternalDistributedSystem system = (InternalDistributedSystem) cache.getDistributedSystem();
-    String redisPassword = system.getConfig().getRedisPassword();
-    final byte[] redisPasswordBytes = Coder.stringToBytes(redisPassword);
-    ServerBootstrap serverBootstrap = new ServerBootstrap();
-
-    serverBootstrap.group(bossGroup, workerGroup).channel(socketClass)
-        .childHandler(createChannelInitializer(redisPasswordBytes))
-        .option(ChannelOption.SO_REUSEADDR, true)
-        .option(ChannelOption.SO_RCVBUF, getBufferSize())
-        .childOption(ChannelOption.SO_KEEPALIVE, true)
-        .childOption(ChannelOption.CONNECT_TIMEOUT_MILLIS, GeodeRedisServer.connectTimeoutMillis)
-        .childOption(ChannelOption.ALLOCATOR, PooledByteBufAllocator.DEFAULT);
-
-    serverChannel = createBoundChannel(serverBootstrap);
-  }
-
-  @SuppressWarnings("deprecation")
-  private Class<? extends ServerChannel> initializeEventLoopGroups() {
-    ThreadFactory selectorThreadFactory =
-        new LoggingThreadFactory("GeodeRedisServer-SelectorThread-", false);
-
-    ThreadFactory workerThreadFactory =
-        new LoggingThreadFactory("GeodeRedisServer-WorkerThread-", true);
-
-    ThreadFactory subscriberThreadFactory =
-        new LoggingThreadFactory("GeodeRedisServer-SubscriberThread-", true);
-
-    Class<? extends ServerChannel> socketClass;
-    if (singleThreadPerConnection) {
-      bossGroup =
-          new io.netty.channel.oio.OioEventLoopGroup(Integer.MAX_VALUE, selectorThreadFactory);
-      workerGroup =
-          new io.netty.channel.oio.OioEventLoopGroup(Integer.MAX_VALUE, workerThreadFactory);
-      socketClass = io.netty.channel.socket.oio.OioServerSocketChannel.class;
-    } else {
-      bossGroup = new NioEventLoopGroup(numSelectorThreads, selectorThreadFactory);
-      workerGroup = new NioEventLoopGroup(numWorkerThreads, workerThreadFactory);
-      subscriberGroup = new NioEventLoopGroup(numWorkerThreads, subscriberThreadFactory);
-      socketClass = NioServerSocketChannel.class;
-    }
-    return socketClass;
-  }
-
-  private Channel createBoundChannel(ServerBootstrap serverBootstrap)
-      throws InterruptedException, UnknownHostException {
-    InetAddress bindAddress = getBindAddress();
-    int port = serverPort == RANDOM_PORT_INDICATOR ? 0 : serverPort;
-    ChannelFuture channelFuture =
-        serverBootstrap.bind(new InetSocketAddress(bindAddress, port)).sync();
-
-    serverPort = ((InetSocketAddress) channelFuture.channel().localAddress()).getPort();
-
-    logStartupMessage(bindAddress);
-
-    return channelFuture.channel();
-  }
-
-  private void logStartupMessage(InetAddress bindAddress) {
-    String logMessage = "GeodeRedisServer started {" + bindAddress + ":" + serverPort
-        + "}, Selector threads: " + numSelectorThreads;
-    if (singleThreadPerConnection) {
-      logMessage += ", One worker thread per connection";
-    } else {
-      logMessage += ", Worker threads: " + numWorkerThreads;
-    }
-    logger.info(logMessage);
-  }
-
-  private ChannelInitializer<SocketChannel> createChannelInitializer(byte[] redisPasswordBytes) {
-    return new ChannelInitializer<SocketChannel>() {
-      @Override
-      public void initChannel(SocketChannel socketChannel) {
-        if (logger.isDebugEnabled()) {
-          logger.debug(
-              "GeodeRedisServer-Connection established with " + socketChannel.remoteAddress());
-        }
-        ChannelPipeline pipeline = socketChannel.pipeline();
-        addSSLIfEnabled(socketChannel, pipeline);
-        pipeline.addLast(ByteToCommandDecoder.class.getSimpleName(), new ByteToCommandDecoder());
-        pipeline.addLast(new WriteTimeoutHandler(10));
-        pipeline.addLast(ExecutionHandlerContext.class.getSimpleName(),
-            new ExecutionHandlerContext(socketChannel, GeodeRedisServer.this,
-                redisPasswordBytes));
-      }
-    };
-  }
-
-  private void addSSLIfEnabled(SocketChannel ch, ChannelPipeline p) {
-
-    SSLConfig sslConfigForComponent =
-        SSLConfigurationFactory.getSSLConfigForComponent(
-            ((InternalDistributedSystem) cache.getDistributedSystem()).getConfig(),
-            SecurableCommunicationChannel.SERVER);
-
-    if (!sslConfigForComponent.isEnabled()) {
-      return;
-    }
-
-    SslContext sslContext;
-    try {
-      KeyStore ks = KeyStore.getInstance("JKS");
-      ks.load(new FileInputStream(sslConfigForComponent.getKeystore()),
-          sslConfigForComponent.getKeystorePassword().toCharArray()/**/);
-
-      // Set up key manager factory to use our key store
-      KeyManagerFactory kmf =
-          KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
-      kmf.init(ks, sslConfigForComponent.getKeystorePassword().toCharArray());
-
-      SslContextBuilder sslContextBuilder = SslContextBuilder.forServer(kmf);
-      sslContext = sslContextBuilder.build();
-
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-    p.addLast(sslContext.newHandler(ch.alloc()));
-  }
-
-  /**
-   * Helper method to get GemFire set socket buffer size, possibly a default of 32k
-   *
-   * @return Buffer size to use for server
-   */
-  private int getBufferSize() {
-    InternalDistributedSystem system = (InternalDistributedSystem) cache.getDistributedSystem();
-    return system.getConfig().getSocketBufferSize();
   }
 
   /**
@@ -412,25 +140,9 @@ public class GeodeRedisServer {
   public synchronized void shutdown() {
     if (!shutdown) {
       logger.info("GeodeRedisServer shutting down");
-      passiveExpirationManager.close();
-      ChannelFuture closeFuture = null;
-      if (serverChannel != null) {
-        closeFuture = serverChannel.closeFuture();
-      }
-      workerGroup.shutdownGracefully();
-      Future<?> bossFuture = bossGroup.shutdownGracefully();
-      if (serverChannel != null) {
-        serverChannel.close();
-      }
-      bossFuture.syncUninterruptibly();
-      if (closeFuture != null) {
-        closeFuture.syncUninterruptibly();
-      }
+      passiveExpirationManager.stop();
+      nettyRedisServer.close();
       shutdown = true;
     }
-  }
-
-  public int getPort() {
-    return serverPort;
   }
 }
