+++ geode-redis/src/main/java/org/apache/geode/redis/internal/executor/server/ShutDownExecutor.java
@@ -25,7 +25,7 @@ public class ShutDownExecutor extends AbstractExecutor {
   @Override
   public RedisResponse executeCommand(Command command,
       ExecutionHandlerContext context) {
-    context.getServer().shutdown();
+    context.shutdown();
     return RedisResponse.nil();
   }
 }
